+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - **[FIX]** Remove DOM Warning on `MainContent` component
 - **[FIX]** Fix markup `Item` tag label by replacing `div` children with `span`
 - **[BREAKING CHANGE]** Import `icons` as default instead of named export for a better code splitting
-  [...]
+- **[UPDATE]** Make `Review` more semantic with proper HTML and Review/Rating microdata
+- [...]
 
 # v25.0.0 (02/04/2020)